# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RefuLearn API",
    "description": "A comprehensive API for the RefuLearn platform supporting refugee education and skill development",
    "version": "1.0.0",
    "contact": {
      "name": "RefuLearn Team",
      "email": "support@refulearn.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.refulearn.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    },
  "paths": {
...
    "/api/peer-learning/groups/{groupId}/comments": {
      "post": {
        "tags": ["peer-learning"],
        "summary": "Add comment to group",
        "description": "Add a new comment to a peer learning group",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "groupId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string", "minLength": 1, "maxLength": 1000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Comment added successfully" },
          "400": { "description": "Invalid input data" }
        }
      }
    },
    "/api/peer-learning/groups/{groupId}/comments/{commentId}/replies": {
      "post": {
        "tags": ["peer-learning"],
        "summary": "Add reply to comment",
        "description": "Add a reply to a specific comment in a peer learning group",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "groupId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "commentId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string", "minLength": 1, "maxLength": 1000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Reply added successfully" },
          "400": { "description": "Invalid input data" }
        }
      }
    }
  }
}